+++ src/ui/RealmBrowser/Content/index.tsx
@@ -125,7 +125,7 @@ export interface IBaseContentContainerProps {
   onClassFocussed?: ClassFocussedHandler;
   onHighlightChange?: (
     highlight: IHighlight | undefined,
-    collection: Realm.Collection<any>,
+    collection: Realm.OrderedCollection<any>,
   ) => void;
   onListFocussed?: ListFocussedHandler;
   onSingleListFocussed?: SingleListFocussedHandler;
@@ -149,7 +149,6 @@ export interface IReadWriteContentContainerProps
   onCancelTransaction: () => void;
   onCommitTransaction: () => void;
   onRealmChanged: () => void;
-  permissionSidebar: boolean;
   readOnly: boolean;
   realm: Realm;
 }
@@ -164,7 +163,6 @@ export interface IContentContainerState {
   error?: Error;
   hideSystemClasses: boolean;
   highlight?: IHighlight;
-  isPermissionSidebarOpen: boolean;
   query: string;
   selectObjectDialog: ISelectObjectDialog;
   sorting?: ISorting;
@@ -180,8 +178,6 @@ class ContentContainer extends React.Component<
     createObjectDialog: { isOpen: false },
     deleteObjectsDialog: { isOpen: false },
     hideSystemClasses: !store.shouldShowSystemClasses(),
-    isPermissionSidebarOpen:
-      !this.props.readOnly && this.props.permissionSidebar,
     query: '',
     selectObjectDialog: { isOpen: false },
   };
@@ -196,7 +192,11 @@ class ContentContainer extends React.Component<
 
   private clickTimeout?: any;
   private filteredSortedResults = memoize(
-    (results: Realm.Collection<any>, query: string, sorting?: ISorting) => {
+    (
+      results: Realm.OrderedCollection<any>,
+      query: string,
+      sorting?: ISorting,
+    ) => {
       let filterError: Error | undefined;
       if (query) {
         try {
@@ -311,7 +311,6 @@ class ContentContainer extends React.Component<
       filteredSortedResults: results,
       focus,
       highlight: this.state.highlight,
-      isPermissionSidebarOpen: this.state.isPermissionSidebarOpen,
       onCellChange: this.onCellChange,
       onCellClick: this.onCellClick,
       onCellHighlighted: this.onCellHighlighted,
@@ -348,9 +347,6 @@ class ContentContainer extends React.Component<
         onAddColumnClick: this.props.onAddColumnClick,
         onCancelTransaction: this.props.onCancelTransaction,
         onCommitTransaction: this.props.onCommitTransaction,
-        onPermissionSidebarToggle: this.props.permissionSidebar
-          ? this.onPermissionSidebarToggle
-          : undefined,
         onReorderingEnd: this.onReorderingEnd,
         onReorderingStart: this.onReorderingStart,
         realm: this.props.realm,
@@ -899,12 +895,6 @@ class ContentContainer extends React.Component<
         : undefined;
 
     this.onShowCreateObjectDialog(className, embeddedInfo);
-  };
-
-  private onPermissionSidebarToggle = () => {
-    this.setState({
-      isPermissionSidebarOpen: !this.state.isPermissionSidebarOpen,
-    });
   };
 
   private onReorderingStart: ReorderingStartHandler = () => {