+++ src/ui/RealmBrowser/RealmBrowser.tsx
@@ -66,7 +66,7 @@ export interface IRealmBrowserProps {
   isPropertyNameAvailable: (name: string) => boolean;
   jsonViewerDialog: null | { value: unknown };
   onAddClass: (schema: Realm.ObjectSchema) => void;
-  onAddProperty: (name: string, type: Realm.PropertyType) => void;
+  onAddProperty: (name: string, type: Realm.PropertyType | string) => void;
   onAddSubscription: (schemaName: string, query: string) => void;
   onCancelTransaction: () => void;
   onClassFocussed: ClassFocussedHandler;
@@ -178,7 +178,6 @@ export const RealmBrowser = ({
             onSingleListFocussed={onSingleListFocussed}
             onRealmChanged={onRealmChanged}
             onShowJsonViewerDialog={onShowJsonViewerDialog}
-            permissionSidebar={true}
             progress={progress}
             readOnly={editMode === EditMode.Disabled}
             realm={realm}